Add valid draft model for existing json metadata
